+++ tksheet/column_headers.py
@@ -56,8 +56,7 @@ from .vars import (
 
 class ColumnHeaders(tk.Canvas):
     def __init__(self, *args, **kwargs):
-        tk.Canvas.__init__(
-            self,
+        super().__init__(
             kwargs["parent"],
             background=kwargs["parent"].ops.header_bg,
             highlightthickness=0,